bookIsbn type 수정

+++ kh7semimain/src/library/servlet/book/BookInsertServlet.java
@@ -18,7 +18,7 @@ public class BookInsertServlet extends HttpServlet {
 		try {
 			req.setCharacterEncoding("UTF-8");
 			BookDto bookDto = new BookDto();
-			bookDto.setBookIsbn(Integer.parseInt(req.getParameter("bookIsbn")));
+			bookDto.setBookIsbn(Long.parseLong(req.getParameter("bookIsbn")));
 			bookDto.setGenreNo(Integer.parseInt(req.getParameter("GenreNo")));
 			bookDto.setBookTitle(req.getParameter("bookName"));
 			bookDto.setBookAuthor(req.getParameter("bookAuthor"));
